[Refactor] 씬에 존재하는 UI, static으로 변경
