+++ JKC_Fallguys/Assets/1_Scripts/Scene/00_Login/LoginSceneInitializer.cs
@@ -13,10 +13,6 @@ public class LoginSceneInitializer : SceneInitializer
     protected override void OnGetResources()
     {
         Instantiate(DataManager.GetGameObjectData
-            (PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.Login, "LoginBackgroundViewController"));
-        Instantiate(DataManager.GetGameObjectData
-            (PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.Login, "LoginPanelViewController"));
-        Instantiate(DataManager.GetGameObjectData
             (PathLiteral.Prefabs, PathLiteral.Manager, "PhotonLoginManager"));
     }
 }
