+++ JKC_Fallguys/Assets/1_Scripts/Scene/04_GameLoading/GameLoadingSceneInitializer.cs
@@ -1,6 +1,5 @@
 using LiteralRepository;
 using Photon.Pun;
-using UnityEngine;
 
 public class GameLoadingSceneInitializer : SceneInitializer
 {
@@ -8,40 +7,12 @@ public class GameLoadingSceneInitializer : SceneInitializer
     {
         Instantiate(DataManager.GetGameObjectData
             (PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.GameLoading, "GameLoadingSceneManager"));
-        Instantiate(DataManager.GetGameObjectData
-            (PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.GameLoading, PathLiteral.UI, "GameLoadingBackgroundImage"));
-        Instantiate(DataManager.GetGameObjectData
-            (PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.GameLoading, PathLiteral.UI, "HorizontalRendererViewController"));
-        Instantiate(DataManager.GetGameObjectData
-            (PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.GameLoading, PathLiteral.UI, "GameLoadingMainPanelViewController"));
-        Instantiate(DataManager.GetGameObjectData
-            (PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.GameLoading, PathLiteral.UI, "WhiteScreenViewController"));
-        Instantiate(DataManager.GetGameObjectData
-            (PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.GameLoading, PathLiteral.UI, "MapInformationViewController"));
 
         if (PhotonNetwork.IsMasterClient)
         {
             string filePath = DataManager.SetDataPath(PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.GameLoading, "MapSelectionManager");
-    
-            GameObject[] allObjects = FindObjectsOfType<GameObject>();
-            foreach (GameObject obj in allObjects)
-            {
-                if (obj.name == "MapSelectionManager(Clone)") 
-                {
-                    return;
-                }
-                
-                if (obj.name == "MapSelectionManager") 
-                {
-                    return;
-                }
-            }
-
-            GameObject newObject = PhotonNetwork.Instantiate(filePath, transform.position, transform.rotation);
-            PhotonView photonView = newObject.GetComponent<PhotonView>();
-            photonView.TransferOwnership(PhotonNetwork.LocalPlayer);
+            PhotonNetwork.Instantiate(filePath, transform.position, transform.rotation);
         }
-
     }
 
     protected override void InitializeModel()
