+++ JKC_Fallguys/Assets/1_Scripts/Scene/05_RoundResult.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: e46a1bd209c56402c8a8c31bb978f091
+guid: 006c6dd0d4fd943419b1784733af849e
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
