+++ JKC_Fallguys/Assets/1_Scripts/UI/03_Stage/ResultInStage/ResultInStagePresenter.cs
@@ -22,6 +22,7 @@ public class ResultInStagePresenter : Presenter
     protected override void OnUpdatedModel()
     {
         int actorNumber = PhotonNetwork.LocalPlayer.ActorNumber;
+        Debug.Log($"ResultInActorNumber : {PhotonNetwork.LocalPlayer.ActorNumber}");
         StageDataManager.Instance.GetCurrentState(actorNumber)
             .Skip(1) // 초기 Default로 설정된것은 Skip.
             .DistinctUntilChanged()
