+++ JKC_Fallguys/Assets/2_Scenes/05_RoundResult.unity
@@ -445,75 +445,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1001 &1122012182
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 5433952250778063412, guid: 7be0e59d2b6b945fd803115abdc32c96,
-        type: 3}
-      propertyPath: m_Name
-      value: PhotonClear
-      objectReference: {fileID: 0}
-    - target: {fileID: 6548120757684889407, guid: 7be0e59d2b6b945fd803115abdc32c96,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 4
-      objectReference: {fileID: 0}
-    - target: {fileID: 6548120757684889407, guid: 7be0e59d2b6b945fd803115abdc32c96,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: -69.05688
-      objectReference: {fileID: 0}
-    - target: {fileID: 6548120757684889407, guid: 7be0e59d2b6b945fd803115abdc32c96,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 10.116463
-      objectReference: {fileID: 0}
-    - target: {fileID: 6548120757684889407, guid: 7be0e59d2b6b945fd803115abdc32c96,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: -8.286463
-      objectReference: {fileID: 0}
-    - target: {fileID: 6548120757684889407, guid: 7be0e59d2b6b945fd803115abdc32c96,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 6548120757684889407, guid: 7be0e59d2b6b945fd803115abdc32c96,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6548120757684889407, guid: 7be0e59d2b6b945fd803115abdc32c96,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6548120757684889407, guid: 7be0e59d2b6b945fd803115abdc32c96,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6548120757684889407, guid: 7be0e59d2b6b945fd803115abdc32c96,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6548120757684889407, guid: 7be0e59d2b6b945fd803115abdc32c96,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6548120757684889407, guid: 7be0e59d2b6b945fd803115abdc32c96,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 7be0e59d2b6b945fd803115abdc32c96, type: 3}
 --- !u!1 &1587317542
 GameObject:
   m_ObjectHideFlags: 0
