--- JKC_Fallguys/Assets/Jinsoo/RoundResultSceneInitializer.cs
@@ -1,25 +1,0 @@
-using LiteralRepository;
-using Photon.Pun;
-
-public class RoundResultSceneInitializer : SceneInitializer
-{
-    protected override void OnGetResources()
-    {
-        Instantiate
-            (DataManager.GetGameObjectData(PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.RoundResult, "RoundResultViewController"));
-        Instantiate
-            (DataManager.GetGameObjectData(PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.RoundResult, "Platform"));
-        Instantiate
-            (DataManager.GetGameObjectData(PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.RoundResult, "BackgroundImage"));
-        Instantiate
-            (DataManager.GetGameObjectData(PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.RoundResult, "ResultRoundSetupManager"));
-        
-        if (PhotonNetwork.IsMasterClient)
-        {
-            string filePath = 
-                DataManager.SetDataPath(PathLiteral.Prefabs, PathLiteral.Scene, PathLiteral.RoundResult, "SceneChanger");
-
-            PhotonNetwork.Instantiate(filePath, transform.position, transform.rotation);
-        }
-    }
-}
