+++ JKC_Fallguys/ProjectSettings/EditorBuildSettings.asset
@@ -12,7 +12,7 @@ EditorBuildSettings:
     path: Assets/2_Scenes/01_Lobby.unity
     guid: 923b410b6813d41e3960d16c429a0824
   - enabled: 1
-    path: Assets/2_Scenes/02_MatchingStandby.unity
+    path: Assets/2_Scenes/02_Matching.unity
     guid: 3f93cd8575b48499b8958efa396b20c7
   - enabled: 1
     path: Assets/2_Scenes/03_Stage.unity
